refactor: 컨트롤러 테스트를 명확하게 분리
